+++ zellij-tile/src/shim.rs
@@ -1,4 +1,4 @@
-use serde::de::DeserializeOwned;
+use serde::{de::DeserializeOwned, Serialize};
 use std::{io, path::Path};
 
 use crate::data::*;
@@ -6,12 +6,12 @@ use crate::data::*;
 // Subscription Handling
 
 pub fn subscribe(event_types: &[EventType]) {
-    println!("{}", serde_json::to_string(event_types).unwrap());
+    object_to_stdout(&event_types);
     unsafe { host_subscribe() };
 }
 
 pub fn unsubscribe(event_types: &[EventType]) {
-    println!("{}", serde_json::to_string(event_types).unwrap());
+    object_to_stdout(&event_types);
     unsafe { host_unsubscribe() };
 }
 
@@ -38,8 +38,12 @@ pub fn get_plugin_ids() -> PluginIds {
 // Host Functions
 
 pub fn open_file(path: &Path) {
-    println!("{}", path.to_string_lossy());
+    object_to_stdout(&path);
     unsafe { host_open_file() };
+}
+
+pub fn set_timeout(secs: f64) {
+    unsafe { host_set_timeout(secs) };
 }
 
 // Internal Functions
@@ -51,6 +55,11 @@ pub fn object_from_stdin<T: DeserializeOwned>() -> T {
     serde_json::from_str(&json).unwrap()
 }
 
+#[doc(hidden)]
+pub fn object_to_stdout(object: &impl Serialize) {
+    println!("{}", serde_json::to_string(object).unwrap());
+}
+
 #[link(wasm_import_module = "zellij")]
 extern "C" {
     fn host_subscribe();
@@ -60,4 +69,5 @@ extern "C" {
     fn host_set_invisible_borders(invisible_borders: i32);
     fn host_get_plugin_ids();
     fn host_open_file();
+    fn host_set_timeout(secs: f64);
 }